Fiddle with namespace to fix Mirage errors
Not completely clear to me what is happening here, but something about
the Ember upgrade (route-recognizer maybe?) has changed how Mirage
resolves URLs. Possibly related to the rootURL change in Ember CLI 2.7.

The fix in this commit is to add a non-empty namespace to Mirage's
config, with a leading slash (to escape the "/1234/tests" rootURL prefix
under testing). Then update the Slack adapter fiddling, and patch up a
brittle unit test.

Conversation in #ec-mirage Slack here:

https://embercommunity.slack.com/archives/ec-mirage/p1469897059000685